refactor(executor): rename mergeHeap to merge
Fold index_merge_heap.go into index_merge.go and rename the K-way merge
function from mergeHeap to merge, since it's no longer heap-based after
the switch to the loser tournament tree. Update test names accordingly.
